Old changes from my old branch https://github.com/jerome077/k-9/tree/Do_not_store_my_password in a new branch with the current version of k9-mail. This change might not be mainstream, as I saw that it doesn't word together with imap push, but for my uses it is enough, so it's more a branch for me.
"I have written a feature to be able to configure an account so that the ingoing and outcoming password won't be stored in the settings. It works like this:

I have added a checkbox "Don't store my password" in the account options.
At start such accounts are disabled (I put a text "[-]" in front of the account name in the account list to identify such accounts easily)
When you open the account, K-9 will ask you for the password(s) but won't store it in the properties, so that in remains in memory only until K-9 is closed. You can also press cancel to access the stored messages without entering a password.
There is also a new menu item in the menu on the account list to forget all sessions passwords without having to exit K-9."